RT Info Attributes Serialization/Deserialization (#7603)
* Init rt info deserialization

* Add RT Info attributes deserialization

* Add RT Info attributes serialization

* Add Serialization and Deserialization tests

* Code style

* Fix AttributeAdapter place

* Extended rt_info section; added support for multi-field attrs; use version for attr serialization/deserialization

* Align attributes and their usage

* Fix LPT

* Add missing #pragma one

* Fix build

* Change version delimiter to have valid xml

* Fix PromitivesPriority

* Fix tests

* Fix test

* Fix IR Reader; remove empty rt_info

* More explicit attribute name and version representation
