+++ inference-engine/src/transformations/include/transformations/rt_info/disable_constant_folding.hpp
@@ -2,6 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
+#pragma once
+
 #include <assert.h>
 #include <functional>
 #include <memory>
@@ -13,7 +15,7 @@
 #include <transformations_visibility.hpp>
 
 
-namespace ngraph {
+namespace ov {
 
 /**
  * @ingroup ie_runtime_attr_api
@@ -25,15 +27,19 @@ public:
 };
 
 TRANSFORMATIONS_API void disable_constant_folding(const std::shared_ptr<Node>& node);
-}  // namespace ngraph
 
-namespace ov {
-extern template class TRANSFORMATIONS_API VariantImpl<ngraph::DisableConstantFolding>;
+TRANSFORMATIONS_API void enable_constant_folding(const std::shared_ptr<Node>& node);
+
+TRANSFORMATIONS_API bool constant_folding_is_disabled(const std::shared_ptr<Node>& node);
+
+extern template class TRANSFORMATIONS_API VariantImpl<DisableConstantFolding>;
 
 template<>
-class TRANSFORMATIONS_API VariantWrapper<ngraph::DisableConstantFolding> : public VariantImpl<ngraph::DisableConstantFolding> {
+class TRANSFORMATIONS_API VariantWrapper<DisableConstantFolding> : public VariantImpl<DisableConstantFolding> {
 public:
-    OPENVINO_RTTI("DISABLED_CONSTANT_FOLDING");
+    OPENVINO_RTTI("disabled_constant_folding", "0");
+
+    VariantWrapper() = default;
 
     VariantWrapper(const value_type &value) : VariantImpl<value_type>(value) {}
