+++ inference-engine/src/transformations/include/transformations/rt_info/nms_selected_indices.hpp
@@ -2,6 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
+#pragma once
+
 #include <assert.h>
 #include <functional>
 #include <memory>
@@ -28,11 +30,9 @@ extern template class TRANSFORMATIONS_API VariantImpl<NmsSelectedIndices>;
 template<>
 class TRANSFORMATIONS_API VariantWrapper<NmsSelectedIndices> : public VariantImpl<NmsSelectedIndices> {
 public:
-    static constexpr VariantTypeInfo type_info{"NMS_SELECTED_INDICES", 0};
+    OPENVINO_RTTI("nms_selected_indices", "0");
 
-    const VariantTypeInfo &get_type_info() const override {
-        return type_info;
-    }
+    VariantWrapper() = default;
 
     VariantWrapper(const value_type &value) : VariantImpl<value_type>(value) {}
