+++ inference-engine/src/transformations/include/transformations/rt_info/primitives_priority_attribute.hpp
@@ -19,7 +19,7 @@
 #include <ngraph/variant.hpp>
 #include <transformations_visibility.hpp>
 
-namespace ngraph {
+namespace ov {
 
 /**
  * @ingroup ie_runtime_attr_api
@@ -31,6 +31,7 @@ private:
     std::string primitives_priority;
 
 public:
+    friend class VariantWrapper<PrimitivesPriority>;
     /**
      * A default constructor
      */
@@ -54,22 +55,22 @@ public:
  */
 TRANSFORMATIONS_API std::string getPrimitivesPriority(const std::shared_ptr<ngraph::Node> & node);
 
-}  // namespace ngraph
-
-namespace ov {
-
-extern template class TRANSFORMATIONS_API VariantImpl<ngraph::PrimitivesPriority>;
+extern template class TRANSFORMATIONS_API VariantImpl<PrimitivesPriority>;
 
 template<>
-class TRANSFORMATIONS_API VariantWrapper<ngraph::PrimitivesPriority> : public VariantImpl<ngraph::PrimitivesPriority> {
+class TRANSFORMATIONS_API VariantWrapper<PrimitivesPriority> : public VariantImpl<PrimitivesPriority> {
 public:
-    OPENVINO_RTTI("VariantWrapper<PrimitivesPriority>");
+    OPENVINO_RTTI("primitives_priority", "0");
+
+    VariantWrapper() = default;
 
     VariantWrapper(const value_type &value) : VariantImpl<value_type>(value) {}
 
     std::shared_ptr<ov::Variant> merge(const ngraph::NodeVector & nodes) override;
 
     std::shared_ptr<ov::Variant> init(const std::shared_ptr<ngraph::Node> & node) override;
+
+    bool visit_attributes(AttributeVisitor & visitor) override;
 };
 
 }  // namespace ov
