+++ inference-engine/src/transformations/include/transformations/rt_info/strides_property.hpp
@@ -9,19 +9,18 @@
 #include <ngraph/variant.hpp>
 #include <transformations_visibility.hpp>
 
-
 namespace ov {
 template <>
 class TRANSFORMATIONS_API VariantWrapper<ngraph::Strides> : public VariantImpl<ngraph::Strides> {
 public:
-    OPENVINO_RTTI("VariantWrapper<Strides>");
-    VariantWrapper(const value_type& value)
-        : VariantImpl<value_type>(value) {
-    }
-};
+    OPENVINO_RTTI("strides", "0");
 
-} // namespace ov
+    VariantWrapper() = default;
+
+    VariantWrapper(const value_type& value) : VariantImpl<value_type>(value) {}
+};
 
 TRANSFORMATIONS_API bool has_strides_prop(const ngraph::Input<ngraph::Node>& node);
 TRANSFORMATIONS_API ngraph::Strides get_strides_prop(const ngraph::Input<ngraph::Node>& node);
 TRANSFORMATIONS_API void insert_strides_prop(ngraph::Input<ngraph::Node>& node, const ngraph::Strides& strides);
+} // namespace ov
