+++ inference-engine/src/transformations/src/transformations/low_precision/disable_convert_constant_folding_on_const_path.cpp
@@ -13,6 +13,7 @@
 #include <ngraph/rt_info.hpp>
 #include <ngraph/pattern/op/wrap_type.hpp>
 #include <ngraph/variant.hpp>
+#include <transformations/rt_info/disable_constant_folding.hpp>
 
 using namespace ngraph;
 
@@ -53,8 +54,7 @@ ngraph::pass::DisableConvertConstantFoldingOnConstPath::DisableConvertConstantFo
         auto child = target_inputs.begin()->get_node();
         if (ov::is_type<ngraph::opset1::Constant>(parent) &&
             (ov::is_type<ngraph::opset1::Subtract>(child) || ov::is_type<ngraph::opset1::Multiply>(child))) {
-            auto& rtInfo = convert->get_rt_info();
-            rtInfo["DISABLED_CONSTANT_FOLDING"] = std::make_shared<VariantWrapper<std::string>>("");
+            ov::disable_constant_folding(convert);
             return true;
         }
