+++ inference-engine/src/transformations/src/transformations/rt_info/fused_names_attribute.cpp
@@ -38,9 +38,9 @@ std::string ngraph::getFusedNames(const std::shared_ptr<ngraph::Node> &node) {
     const auto &rtInfo = node->get_rt_info();
     using FusedNamesWrapper = VariantWrapper<FusedNames>;
 
-    if (!rtInfo.count(FusedNamesWrapper::get_type_info_static().name)) return {};
+    if (!rtInfo.count(FusedNamesWrapper::get_type_info_static())) return {};
 
-    const auto &attr = rtInfo.at(FusedNamesWrapper::get_type_info_static().name);
+    const auto &attr = rtInfo.at(FusedNamesWrapper::get_type_info_static());
     FusedNames fusedNames = ov::as_type_ptr<FusedNamesWrapper>(attr)->get();
     return fusedNames.getNames();
 }
@@ -51,24 +51,23 @@ std::vector<std::string> ngraph::getFusedNamesVector(const std::shared_ptr<ngrap
     const auto &rtInfo = node->get_rt_info();
     using FusedNamesWrapper = VariantWrapper<FusedNames>;
 
-    if (!rtInfo.count(FusedNamesWrapper::get_type_info_static().name)) return {};
+    if (!rtInfo.count(FusedNamesWrapper::get_type_info_static())) return {};
 
-    const auto &attr = rtInfo.at(FusedNamesWrapper::get_type_info_static().name);
+    const auto &attr = rtInfo.at(FusedNamesWrapper::get_type_info_static());
     FusedNames fusedNames = ov::as_type_ptr<FusedNamesWrapper>(attr)->get();
     return fusedNames.getVectorNames();
 }
 
 template class ov::VariantImpl<FusedNames>;
 
-BWDCMP_RTTI_DEFINITION(VariantWrapper<FusedNames>);
-
 std::shared_ptr<ngraph::Variant> VariantWrapper<FusedNames>::merge(const ngraph::NodeVector & nodes) {
     FusedNames mergedNames;
     for (auto &node : nodes) {
         const auto &rtInfo = node->get_rt_info();
-        if (!rtInfo.count(VariantWrapper<FusedNames>::get_type_info_static().name)) continue;
 
-        const auto attr = rtInfo.at(VariantWrapper<FusedNames>::get_type_info_static().name);
+        if (!rtInfo.count(VariantWrapper<FusedNames>::get_type_info_static())) continue;
+
+        const auto attr = rtInfo.at(VariantWrapper<FusedNames>::get_type_info_static());
         if (auto fusedNames = std::dynamic_pointer_cast<VariantWrapper<FusedNames> >(attr)) {
             mergedNames.fuseWith(fusedNames->get());
         }
@@ -78,4 +77,9 @@ std::shared_ptr<ngraph::Variant> VariantWrapper<FusedNames>::merge(const ngraph:
 
 std::shared_ptr<ngraph::Variant> VariantWrapper<FusedNames>::init(const std::shared_ptr<ngraph::Node> & node) {
     return std::make_shared<VariantWrapper<FusedNames> > (FusedNames(node->get_friendly_name()));
+}
+
+bool VariantWrapper<FusedNames>::visit_attributes(AttributeVisitor &visitor) {
+    visitor.on_attribute("value", m_value.fused_names);
+    return true;
 }
