+++ inference-engine/src/transformations/src/transformations/rt_info/nms_selected_indices.cpp
@@ -6,14 +6,12 @@
 
 template class ov::VariantImpl<ov::NmsSelectedIndices>;
 
-constexpr ov::VariantTypeInfo ov::VariantWrapper<ov::NmsSelectedIndices>::type_info;
-
 void ov::set_nms_selected_indices(Node * node) {
     auto & rt_info = node->get_rt_info();
-    rt_info[VariantWrapper<NmsSelectedIndices>::type_info.name] = make_variant<NmsSelectedIndices>({});
+    rt_info[VariantWrapper<NmsSelectedIndices>::get_type_info_static()] = make_variant<NmsSelectedIndices>({});
 }
 
 bool ov::has_nms_selected_indices(const Node * node) {
     const auto & rt_info = node->get_rt_info();
-    return rt_info.count(VariantWrapper<NmsSelectedIndices>::type_info.name);
+    return rt_info.count(VariantWrapper<NmsSelectedIndices>::get_type_info_static());
 }
