+++ inference-engine/src/transformations/src/transformations/rt_info/primitives_priority_attribute.cpp
@@ -23,13 +23,13 @@ std::string PrimitivesPriority::getPrimitivesPriority() const {
     return primitives_priority;
 }
 
-std::string ngraph::getPrimitivesPriority(const std::shared_ptr<ngraph::Node> &node) {
+std::string ov::getPrimitivesPriority(const std::shared_ptr<ngraph::Node> &node) {
     const auto &rtInfo = node->get_rt_info();
     using PrimitivesPriorityWrapper = VariantWrapper<PrimitivesPriority>;
 
-    if (!rtInfo.count(PrimitivesPriorityWrapper::get_type_info_static().name)) return "";
+    if (!rtInfo.count(PrimitivesPriorityWrapper::get_type_info_static())) return "";
 
-    const auto &attr = rtInfo.at(PrimitivesPriorityWrapper::get_type_info_static().name);
+    const auto &attr = rtInfo.at(PrimitivesPriorityWrapper::get_type_info_static());
     PrimitivesPriority pp = ov::as_type_ptr<PrimitivesPriorityWrapper>(attr)->get();
     return pp.getPrimitivesPriority();
 }
@@ -59,7 +59,7 @@ std::shared_ptr<ngraph::Variant> VariantWrapper<PrimitivesPriority>::merge(const
     }
 
     if (unique_pp.size() > 1) {
-        throw ngraph_error(std::string(get_type_info().name) + " no rule defined for multiple values.");
+        throw ngraph_error(std::string(get_type_info()) + " no rule defined for multiple values.");
     }
 
     std::string final_primitives_priority;
@@ -70,5 +70,10 @@ std::shared_ptr<ngraph::Variant> VariantWrapper<PrimitivesPriority>::merge(const
 }
 
 std::shared_ptr<ngraph::Variant> VariantWrapper<PrimitivesPriority>::init(const std::shared_ptr<ngraph::Node> & node) {
-    throw ngraph_error(std::string(get_type_info().name) + " has no default initialization.");
+    throw ngraph_error(std::string(get_type_info()) + " has no default initialization.");
+}
+
+bool VariantWrapper<PrimitivesPriority>::visit_attributes(AttributeVisitor &visitor) {
+    visitor.on_attribute("value", m_value.primitives_priority);
+    return true;
 }
