+++ inference-engine/src/transformations/src/transformations/serialize.cpp
@@ -18,6 +18,7 @@
 #include "ngraph_ops/type_relaxed.hpp"
 #include "pugixml.hpp"
 #include "transformations/serialize.hpp"
+#include "transformations/rt_info/attributes.hpp"
 
 using namespace ngraph;
 
@@ -175,6 +176,82 @@ private:
     pugi::xml_node& m_xml_node;
 };
 
+class RTInfoSerializer : public ngraph::AttributeVisitor {
+    pugi::xml_node m_node;
+
+public:
+    RTInfoSerializer(const pugi::xml_node node) : m_node(node) {}
+
+    void on_adapter(const std::string &name, ngraph::ValueAccessor<void> &adapter) override {
+        check_attribute_name(name);
+        if (auto a = ngraph::as_type<ngraph::AttributeAdapter<std::set<std::string>>>(&adapter)) {
+            const auto & value = join(a->get());
+            m_node.append_attribute(name.c_str()).set_value(value.c_str());
+        } else {
+            throw ngraph_error("Unsupported attribute type for serialization: " + name);
+        }
+    }
+
+    void on_adapter(const std::string &name, ngraph::ValueAccessor<bool> &adapter) override {
+        check_attribute_name(name);
+        m_node.append_attribute(name.c_str()).set_value(adapter.get());
+    }
+
+    void on_adapter(const std::string &name, ngraph::ValueAccessor<std::string> &adapter) override {
+        check_attribute_name(name);
+        m_node.append_attribute(name.c_str()).set_value(adapter.get().c_str());
+    }
+
+    void on_adapter(const std::string &name, ngraph::ValueAccessor<int64_t> &adapter) override {
+        check_attribute_name(name);
+        m_node.append_attribute(name.c_str()).set_value(adapter.get());
+    }
+
+    void on_adapter(const std::string &name, ngraph::ValueAccessor<double> &adapter) override {
+        check_attribute_name(name);
+        m_node.append_attribute(name.c_str()).set_value(adapter.get());
+    }
+
+    void on_adapter(const std::string &name, ngraph::ValueAccessor<std::vector<int>> &adapter) override {
+        check_attribute_name(name);
+        const auto & value = join(adapter.get());
+        m_node.append_attribute(name.c_str()).set_value(value.c_str());
+    }
+
+    void on_adapter(const std::string &name, ngraph::ValueAccessor<std::vector<int64_t>> &adapter) override {
+        check_attribute_name(name);
+        const auto & value = join(adapter.get());
+        m_node.append_attribute(name.c_str()).set_value(value.c_str());
+    }
+
+    void on_adapter(const std::string &name, ngraph::ValueAccessor<std::vector<uint64_t>> &adapter) override {
+        check_attribute_name(name);
+        const auto & value = join(adapter.get());
+        m_node.append_attribute(name.c_str()).set_value(value.c_str());
+    }
+
+    void on_adapter(const std::string &name, ngraph::ValueAccessor<std::vector<float>> &adapter) override {
+        check_attribute_name(name);
+        const auto & value = join(adapter.get());
+        m_node.append_attribute(name.c_str()).set_value(value.c_str());
+    }
+
+    void on_adapter(const std::string &name, ngraph::ValueAccessor<std::vector<std::string>> &adapter) override {
+        check_attribute_name(name);
+        const auto & value = join(adapter.get());
+        m_node.append_attribute(name.c_str()).set_value(value.c_str());
+    }
+
+    void on_adapter(const std::string &name, ngraph::ValueAccessor<std::shared_ptr<Function>> &adapter) override {
+        throw ngraph_error("Function type is unsupported for rt info serialization");
+    }
+
+    void check_attribute_name(const std::string & name) const {
+        if (name == "name" || name == "version") {
+            throw ngraph_error("Attribute key with name: " + name + " is not allowed. Please use another name");
+        }
+    }
+};
 } // namespace rt_info
 
 class XmlSerializer : public ngraph::AttributeVisitor {
@@ -724,6 +801,27 @@ void ngfunction_2_irv10(pugi::xml_node& netXml,
         // <layers/data> general attributes
         pugi::xml_node data = layer.append_child("data");
 
+        auto append_runtime_info = [](pugi::xml_node & node, const RTMap& attributes) {
+            pugi::xml_node rt_node = node.append_child("rt_info");
+            bool has_attrs = false;
+            for (const auto &item : attributes) {
+                auto attribute_node = rt_node.append_child("attribute");
+                attribute_node.append_attribute("name").set_value(item.second->get_type_info().name);
+                attribute_node.append_attribute("version").set_value(item.second->get_type_info().get_version().c_str());
+                rt_info::RTInfoSerializer serializer(attribute_node);
+                if (!item.second->visit_attributes(serializer)) {
+                    rt_node.remove_child(attribute_node);
+                } else {
+                    has_attrs = true;
+                }
+            }
+            if (!has_attrs) {
+                node.remove_child(rt_node);
+            }
+        };
+
+        append_runtime_info(layer, node->get_rt_info());
+
         int port_id = 0;
         // <layers/input>
         if (node->get_input_size() > 0) {
@@ -748,6 +846,7 @@ void ngfunction_2_irv10(pugi::xml_node& netXml,
                                 .set_value(std::to_string(d.get_length()).c_str());
                     }
                 }
+                append_runtime_info(port, i.get_rt_info());
             }
 
             if (node_type_name == "TensorIterator" || node_type_name == "Loop") {
@@ -798,6 +897,7 @@ void ngfunction_2_irv10(pugi::xml_node& netXml,
                                 .set_value(std::to_string(d.get_length()).c_str());
                     }
                 }
+                append_runtime_info(port, o.get_rt_info());
             }
             if (node_type_name == "TensorIterator" || node_type_name == "Loop") {
                 layer.insert_move_after(output, layer.first_child());
@@ -868,9 +968,6 @@ std::string provide_bin_path(const std::string &xmlPath, const std::string &binP
 }  // namespace
 
 namespace ngraph {
-
-// ! [function_pass:serialize_cpp]
-// serialize.cpp
 bool pass::Serialize::run_on_function(std::shared_ptr<ngraph::Function> f) {
     RUN_ON_FUNCTION_SCOPE(Serialize);
 
@@ -1014,5 +1111,4 @@ bool ngraph::pass::StreamSerialize::run_on_function(std::shared_ptr<ngraph::Func
     // Return false because we didn't change nGraph Function
     return false;
 }
-// ! [function_pass:serialize_cpp]
 }  // namespace ngraph
