+++ inference-engine/tests/functional/inference_engine/lp_transformations/convolution_qdq_transformation.cpp
@@ -174,7 +174,7 @@ const std::vector<ConvolutionQDqTransformationTestValues> testValues = {
             },
             {
                 {},
-                { { 127.f }, ngraph::element::f32, { 6, 1, 1, 1 }, false, 1ul, element::i8, false, { "DISABLED_CONSTANT_FOLDING" } },
+                { { 127.f }, ngraph::element::f32, { 6, 1, 1, 1 }, false, 1ul, element::i8, false, { "disabled_constant_folding_0" } },
                 {}
             },
             { std::vector<float>{ 100.f }, ngraph::element::i8},
@@ -349,7 +349,7 @@ const std::vector<ConvolutionQDqTransformationTestValues> testValues = {
             },
             {
                 {},
-                { { 127.f }, ngraph::element::f32, { 6, 1, 1, 1 }, false, 1ul, element::i8, false, { "DISABLED_CONSTANT_FOLDING" } },
+                { { 127.f }, ngraph::element::f32, { 6, 1, 1, 1 }, false, 1ul, element::i8, false, { "disabled_constant_folding_0" } },
                 {}
             },
             { std::vector<float>{ 2.f }, ngraph::element::i8},
@@ -417,7 +417,7 @@ const std::vector<ConvolutionQDqTransformationTestValues> testValues = {
             },
             {
                 {},
-                { { 127.f }, ngraph::element::f32, { 6, 1, 1, 1 }, false, 1ul, element::i8, false, { "DISABLED_CONSTANT_FOLDING" } },
+                { { 127.f }, ngraph::element::f32, { 6, 1, 1, 1 }, false, 1ul, element::i8, false, { "disabled_constant_folding_0" } },
                 {}
             },
             { std::vector<float>{ 2.f }, ngraph::element::i8},
