+++ inference-engine/tests/unit/inference_engine/ie_compilation_context_test.cpp
@@ -225,24 +225,24 @@ TEST(NetworkContext_CNNNetwork, HashWithPrimitivesPriority) {
 
 TEST(NetworkContext_CNNNetwork, HashWithFusedNames) {
     auto setFusedEmpty = [&](Node::RTMap& rtInfo) {
-        rtInfo[VariantWrapper<FusedNames>::get_type_info_static().name] =
-                std::make_shared<VariantWrapper<FusedNames>>(FusedNames());
+        rtInfo[VariantWrapper<ngraph::FusedNames>::get_type_info_static()] =
+                std::make_shared<VariantWrapper<ngraph::FusedNames>>(ngraph::FusedNames());
     };
     auto setFused = [&](Node::RTMap& rtInfo, const std::string& name) {
-        rtInfo[VariantWrapper<FusedNames>::get_type_info_static().name] =
-                std::make_shared<VariantWrapper<FusedNames>>(FusedNames(name));
+        rtInfo[VariantWrapper<ngraph::FusedNames>::get_type_info_static()] =
+                std::make_shared<VariantWrapper<ngraph::FusedNames>>(ngraph::FusedNames(name));
     };
     checkCustomRt(setFusedEmpty, setFused);
 }
 
 TEST(NetworkContext_CNNNetwork, HashWithPrimitivesPriorityType) {
     auto setPrimEmpty = [&](Node::RTMap& rtInfo) {
-        rtInfo[VariantWrapper<PrimitivesPriority>::get_type_info_static().name] =
-                std::make_shared<VariantWrapper<PrimitivesPriority>>(PrimitivesPriority());
+        rtInfo[VariantWrapper<ov::PrimitivesPriority>::get_type_info_static()] =
+                std::make_shared<VariantWrapper<ov::PrimitivesPriority>>(ov::PrimitivesPriority());
     };
     auto setPrim = [&](Node::RTMap& rtInfo, const std::string& name) {
-        rtInfo[VariantWrapper<PrimitivesPriority>::get_type_info_static().name] =
-                std::make_shared<VariantWrapper<PrimitivesPriority>>(PrimitivesPriority(name));
+        rtInfo[VariantWrapper<ov::PrimitivesPriority>::get_type_info_static()] =
+                std::make_shared<VariantWrapper<ov::PrimitivesPriority>>(ov::PrimitivesPriority(name));
     };
     checkCustomRt(setPrimEmpty, setPrim);
 }
