+++ ngraph/core/src/op/shape_of.cpp
@@ -174,7 +174,7 @@ bool op::v3::ShapeOf::evaluate_upper(const HostTensorVector& output_values) cons
 
 bool op::v3::ShapeOf::constant_fold(OutputVector& output_values, const OutputVector& input_values) {
     OV_ITT_SCOPED_TASK(ov::itt::domains::nGraph, "op::v3::ShapeOf::constant_fold");
-    if (get_rt_info().count("DISABLED_CONSTANT_FOLDING"))
+    if (get_rt_info().count("disabled_constant_folding_0"))
         return false;
     return shape_of::constant_fold_shape_of(this, output_values[0], input_values[0]);
 }
@@ -233,7 +233,7 @@ bool op::v0::ShapeOf::has_evaluate() const {
 
 bool op::v0::ShapeOf::constant_fold(OutputVector& output_values, const OutputVector& input_values) {
     OV_ITT_SCOPED_TASK(ov::itt::domains::nGraph, "op::v0::ShapeOf::constant_fold");
-    if (get_rt_info().count("DISABLED_CONSTANT_FOLDING"))
+    if (get_rt_info().count("disabled_constant_folding_0"))
         return false;
     return shape_of::constant_fold_shape_of(this, output_values[0], input_values[0]);
 }
